fix: use -l flag instead of -c and cacheFile
This is because the script implementation had a cache file shared
by all repositories, thus with mixed commits. I could have found a
way to make it work, but I chose to just use the -l flag: it will
decrease the size of the log.html file too, which would grow as
the number of commits increases.

Signed-off-by: Francesco Saccone <francesco@francescosaccone.com>