fix: remove all sendmail flags
'-v' is not really needed. -F and -f are not needed as well
because the From: and To: values are already specified in the
email. It should also fix a bug where the email receives a response
from the email daemon of failure to send the email to the unkown
"Saccone@hermes.mail.francescosaccone.com" address, supposedly
because of the wrong recipient argument "Saccone" from the
"Francesco Saccone" full name.

Signed-off-by: Francesco Saccone <francesco@francescosaccone.com>